fix: moving some helpers at parent level
